feat(components/table): extract xm-table directive
